2054: Refactor
