+++ accessibility_monitoring_platform/apps/simplified/utils.py
@@ -13,15 +13,21 @@ from django.contrib.auth.models import User
 from django.db import models
 from django.db.models import Case as DjangoCase
 from django.db.models import Q, QuerySet, When
+from django.http import StreamingHttpResponse
 from django.urls import reverse
 
 from ..audits.models import Audit
+from ..cases.csv_export import csv_output_generator
 from ..common.form_extract_utils import (
     FieldLabelAndValue,
     extract_form_labels_and_values,
 )
 from ..common.sitemap import PlatformPage, PlatformPageGroup, Sitemap
 from ..common.utils import build_filters, diff_model_fields
+from .csv_export import (
+    SIMPLIFIED_CASE_COLUMNS_FOR_EXPORT,
+    SIMPLIFIED_FEEDBACK_SURVEY_COLUMNS_FOR_EXPORT,
+)
 from .models import (
     COMPLIANCE_FIELDS,
     CaseCompliance,
@@ -342,3 +348,35 @@ def record_simplified_model_create_event(
         event_type=SimplifiedEventHistory.Type.CREATE,
         difference=json.dumps(model_object_fields, default=str),
     )
+
+
+def download_simplified_cases(
+    simplified_cases: QuerySet[SimplifiedCase], filename: str = "simplified_cases.csv"
+) -> StreamingHttpResponse:
+    """Given a Case queryset, download the data in csv format"""
+
+    response = StreamingHttpResponse(
+        csv_output_generator(
+            cases=simplified_cases,
+            columns_for_export=SIMPLIFIED_CASE_COLUMNS_FOR_EXPORT,
+        ),
+        content_type="text/csv",
+    )
+    response["Content-Disposition"] = f"attachment; filename={filename}"
+    return response
+
+
+def download_simplified_feedback_survey_cases(
+    cases: QuerySet[SimplifiedCase],
+    filename: str = "simplified_feedback_survey_cases.csv",
+) -> StreamingHttpResponse:
+    """Given a Case queryset, download the feedback survey data in csv format"""
+    response = StreamingHttpResponse(
+        csv_output_generator(
+            cases=cases,
+            columns_for_export=SIMPLIFIED_FEEDBACK_SURVEY_COLUMNS_FOR_EXPORT,
+        ),
+        content_type="text/csv",
+    )
+    response["Content-Disposition"] = f"attachment; filename={filename}"
+    return response
